% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.kdeAlgo.R
\name{print.kdeAlgo}
\alias{print.kdeAlgo}
\title{Prints a kdeAlgo Object}
\usage{
\method{print}{kdeAlgo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"kdeAlgo"}}

\item{...}{optional arguments passed to generic function}
}
\description{
Basic information of a kdeAlgo object is printed.
}
\seealso{
\code{\link{kdeAlgoObject}}, \code{\link{kdeAlgo}}
}
