% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sem.R
\name{print.sem}
\alias{print.sem}
\title{Prints a sem Object}
\usage{
\method{print}{sem}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"sem"}.}

\item{...}{optional arguments passed to generic function}
}
\description{
Basic information of a sem object is printed
}
\seealso{
\code{\link{semObject}}, \code{\link{semLm}}, \code{\link{semLme}}
}
