% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sem.R
\name{print.summary.sem}
\alias{print.summary.sem}
\title{Prints a summary.sem Object}
\usage{
print.summary.sem(x, ...)
}
\arguments{
\item{x}{an object of class "summary.sem".}

\item{...}{additional arguments that are not used in this method.}
}
\description{
The elements described in summary.sem are printed.
}
