% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kdeAlgo.R
\name{plot.kdeAlgo}
\alias{plot.kdeAlgo}
\title{Plot Diagnostics for a kdeAlgo Object}
\usage{
\method{plot}{kdeAlgo}(x, indicator = NULL, ...)
}
\arguments{
\item{x}{an object of type "kdeAlgo", typical result of \code{\link{kdeAlgo}}}

\item{indicator}{a vector of indicator names specifying for which indicators
convergence plots are plotted, e.g. \code{c("mean", "gini")}}

\item{...}{optional arguments passed to generic function.}
}
\value{
Convergence and density plots.
}
\description{
Plots the esimtated density from the interval censored variable.
Also, convergence plots are given for all estimated statistical indicators.
The estimated indicator is plotted against the iteration step of the KDE-Algorithm.
A vertical line indicates the end of the burn-in period.
}
\seealso{
\code{\link{kdeAlgoObject}}, \code{\link{kdeAlgo}}
}
